Revert binary backward-incompatible changes
 * bf19ad6: Make `Mock.Object` non-virtual
 * ccf9a02: Remove yet another superfluous extension method class
 * 49a2ccf: Rename extension method for `mock.Protected()`
 * 41ef023: Put async extension methods in `AsyncExtensions`
 * 8aacd8b: Remove superfluous extension methods
 * f605bfe: Fix name of class gen'd by `ReturnsExtensions.tt`
 * 6fadc8f: Merge `ObsoleteMockExtensions` into `MockExtensions`
 * 0cbdd87: Convert `Expect*` methods into extension methods
 * 0669d43: Remove redundant class `MockLegacyExtensions`
 * 85e0e7d: `MockFactory` <--> `MockRepository`

Those would be completely fine for people relying on Moq directly, as
they could simply recompile their code. However, for anyone using Moq
through a 3rd-party library that was compiled against an earlier ver-
sion, they break user code due to binary incompatibility. Forcing all
3rd-party libs to recompile and republish without protecting them with
a semver major version increase is perhaps asking a bit much. :)
