+++ src/Moq/AsInterface.cs
@@ -50,6 +50,11 @@ namespace Moq
 
 		internal override Type[] InheritedInterfaces => this.owner.InheritedInterfaces;
 
+		public override TInterface Object
+		{
+			get { return this.owner.Object as TInterface; }
+		}
+
 		internal override SetupCollection Setups => this.owner.Setups;
 
 		public override Switches Switches
@@ -65,7 +70,10 @@ namespace Moq
 			return this.owner.As<TNewInterface>();
 		}
 
-		protected internal override object OnGetObject() => this.owner.OnGetObject();
+		protected override object OnGetObject()
+		{
+			return this.owner.Object;
+		}
 
 		public override string ToString()
 		{
