+++ src/Moq/ReturnsExtensions.Generated.cs
@@ -2,6 +2,7 @@
 // All rights reserved. Licensed under the BSD 3-Clause License; see License.txt.
 
 using System;
+using System.ComponentModel;
 using System.Threading.Tasks;
 
 using Moq.Language;
@@ -9,7 +10,11 @@ using Moq.Language.Flow;
 
 namespace Moq
 {
-	public static partial class AsyncExtensions
+	/// <summary>
+	/// Defines async extension methods on IReturns.
+	/// </summary>
+	[EditorBrowsable(EditorBrowsableState.Never)]
+	public static class GeneratedReturnsExtensions
 	{
 		/// <summary>
 		/// Specifies a function that will calculate the value to return from the asynchronous method.
