+++ tests/Moq.Tests/ProtectedMockFixture.cs
@@ -14,7 +14,7 @@ namespace Moq.Tests
 		[Fact]
 		public void ThrowsIfNullMock()
 		{
-			Assert.Throws<ArgumentNullException>(() => ((Mock<string>)null).Protected());
+			Assert.Throws<ArgumentNullException>(() => ProtectedExtension.Protected((Mock<string>)null));
 		}
 
 		[Fact]
